Random powerups granted
Used Joey's timer to randomly grant power ups:
current bug if you win and you have god mode it lags out? Only until
next game starts so it's not a huge issue? but maybe it is?
Also vision is the most common power up granted, we should not do that.
It was solely for debugging purposes.